Fixed player selection
I fixed my character selection on my home menu and made that selection appear in the game. I had to delete and create a bunch of prefabs, and that was a whole headache. I also fixed some of the audio for the buttons to play when clicked. I tried to do a score but I could not even get it to show up on the scene.